requirements.txt
flask==3.0.0
requests==2.32.3
python-whois==0.8.0
dnspython==2.6.1